requirements.txt
Pillow
streamlit
pandas
google-cloud-storage==2.6.0
gcsfs==0.6.2